lambdas/requirements.txt
nanoid==2.0.0
pytest==7.4.4
python-dotenv==1.0.0